requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=1.4.0=pypi_0
addict=2.4.0=pypi_0
asttokens=2.2.1=pypi_0
backcall=0.2.0=pypi_0
bzip2=1.0.8=h7f98852_4
ca-certificates=2023.5.7=hbcca054_0
cachetools=5.3.1=pypi_0
certifi=2023.5.7=pypi_0
charset-normalizer=3.1.0=pypi_0
click=8.1.6=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.0.7=pypi_0
cycler=0.11.0=pypi_0
decorator=5.1.1=pypi_0
executing=1.2.0=pypi_0
fonttools=4.39.4=pypi_0
google-auth=2.20.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grad-cam=1.4.8=pypi_0
grpcio=1.54.2=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.7.0=pypi_0
importlib-resources=5.12.0=pypi_0
ipython=8.12.2=pypi_0
jedi=0.18.2=pypi_0
joblib=1.3.1=pypi_0
kiwisolver=1.4.4=pypi_0
ld_impl_linux-64=2.40=h41732ed_0
libffi=3.4.2=h7f98852_5
libgcc-ng=12.2.0=h65d4601_19
libgomp=12.2.0=h65d4601_19
libnsl=2.0.0=h7f98852_0
libsqlite=3.42.0=h2797004_0
libuuid=2.38.1=h0b41bf4_0
libzlib=1.2.13=h166bdaf_4
markdown=3.4.3=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.7.1=pypi_0
matplotlib-inline=0.1.6=pypi_0
mdurl=0.1.2=pypi_0
mmcv-full=1.7.1=pypi_0
mmdet=3.1.0=pypi_0
model-index=0.1.11=pypi_0
ncurses=6.3=h27087fc_1
numpy=1.24.3=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.7.0.72=pypi_0
opendatalab=0.0.9=pypi_0
openmim=0.3.9=pypi_0
openssl=3.1.1=hd590300_0
ordered-set=4.1.0=pypi_0
packaging=23.1=pypi_0
pandas=2.0.1=pypi_0
parso=0.8.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.5.0=pypi_0
pip=23.1.2=pyhd8ed1ab_0
platformdirs=3.9.1=pypi_0
prompt-toolkit=3.0.38=pypi_0
protobuf=4.23.3=pypi_0
psutil=5.9.5=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycocotools=2.0.6=pypi_0
pycryptodome=3.18.0=pypi_0
pygments=2.15.1=pypi_0
pyparsing=3.0.9=pypi_0
python=3.8.16=he550d4f_1_cpython
python-dateutil=2.8.2=pypi_0
pytz=2023.3=pypi_0
pyyaml=6.0=pypi_0
readline=8.2=h8228510_1
requests=2.30.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rich=13.4.2=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.3.0=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.12.2=pypi_0
setuptools=67.7.2=pyhd8ed1ab_0
shapely=2.0.1=pypi_0
six=1.16.0=pypi_0
stack-data=0.6.2=pypi_0
tabulate=0.9.0=pypi_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.1=pypi_0
terminaltables=3.1.10=pypi_0
thop=0.1.1-2209072238=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h27826a3_0
tomli=2.0.1=pypi_0
torch=1.12.1+cu102=pypi_0
torchaudio=0.12.1+cu102=pypi_0
torchvision=0.13.1+cu102=pypi_0
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pypi_0
ttach=0.0.3=pypi_0
typing-extensions=4.5.0=pypi_0
tzdata=2023.3=pypi_0
urllib3=1.26.16=pypi_0
wcwidth=0.2.6=pypi_0
werkzeug=2.3.6=pypi_0
wheel=0.40.0=pyhd8ed1ab_0
xz=5.2.6=h166bdaf_0
yapf=0.40.1=pypi_0
zipp=3.15.0=pypi_0
